refactor: move components between app and theme
