+++ rpsych-cohend/src/components/settings/VizSettings.js
@@ -5,12 +5,14 @@ import Container from "@material-ui/core/Container";
 import SettingsInput from "gatsby-theme-rpsych-viz/src/components/SettingsInput";
 import SaveButton from "gatsby-theme-rpsych-viz/src/components/SaveButton";
 import { Typography } from "@material-ui/core";
-import { SettingsContext } from "../../App";
-import { makeStyles } from "@material-ui/core/styles";
-import { SketchPicker } from "react-color";
+import { SettingsContext } from "../../Viz";
+import { makeStyles, useTheme } from "@material-ui/core/styles";
+import useMediaQuery from '@material-ui/core/useMediaQuery';
+import { SketchPicker, HuePicker } from "react-color";
 import reactCSS from "reactcss";
 import FormControlLabel from '@material-ui/core/FormControlLabel';
 import Switch from '@material-ui/core/Switch';
+import DarkModeToggle from 'gatsby-theme-rpsych/src/components/DarkModeToggle'
 
 const useStyles = makeStyles(theme => ({
   container: {
@@ -20,20 +22,13 @@ const useStyles = makeStyles(theme => ({
   }
 }));
 
-const ColorPicker = ({ dist }) => {
-  const { state, dispatch } = useContext(SettingsContext);
-  const [color, setColor] = useState(state[`color${dist}`]);
+const DesktopColor = ({color, handleChange}) => {
+  const theme = useTheme();
   const [toggle, setToggle] = useState(false);
-
   const handleClick = () => setToggle(!toggle);
   const handleClose = () => setToggle(false);
-  const handleChange = ({ rgb }) => {
-    setColor(rgb);
-    dispatch({ name: `color${dist}`, value: rgb });
-  };
-
   const styles = reactCSS({
-    default: {
+    'default': {
       color: {
         width: "36px",
         height: "14px",
@@ -42,29 +37,30 @@ const ColorPicker = ({ dist }) => {
       },
       swatch: {
         padding: "5px",
-        background: state.darkMode ? '#272727':'#fff',
+        background: theme.palette.type === 'dark' ? '#272727':'#fff',
         borderRadius: "1px",
         boxShadow: "0 0 0 1px rgba(0,0,0,.1)",
         display: "inline-block",
         cursor: "pointer"
       },
       popover: {
-        backgroundColor: "red",
+        backgroundColor: "green",
         position: "absolute",
         left:"10px",
-        zIndex: "2",
+        touchAction: "none",
+        zIndex: 2
       },
       cover: {
         position: "fixed",
         top: "0px",
         right: "0px",
         bottom: "0px",
-        left: "0px"
+        left: "0px",
       },
-      }
+      },
   });
   return (
-    <div>
+    <>
       <div style={styles.swatch} onClick={handleClick}>
         <div style={styles.color} />
       </div>
@@ -74,7 +70,29 @@ const ColorPicker = ({ dist }) => {
           <SketchPicker color={color} onChange={handleChange} />
         </div>
       ) : null}
+    </>
+  );
+  }
+
+const ColorPicker = ({ dist }) => {
+  const { state, dispatch } = useContext(SettingsContext);
+  const [color, setColor] = useState(state[`color${dist}`]);
+  const theme = useTheme();
+  const matches = useMediaQuery(theme.breakpoints.down("xs"));
+
+  const handleChange = ({ rgb }) => {
+    console.log(rgb)
+    setColor(rgb);
+    dispatch({ name: `color${dist}`, value: rgb });
+  };
+  const mobile = useMediaQuery(theme.breakpoints.down("xs"));
+
+  return mobile ? (
+    <div style={{touchAction: "none"}}>
+    <HuePicker color={color} onChange={handleChange} width="100%" />
     </div>
+  ) : (
+    <DesktopColor color={color} handleChange={handleChange} state={state}/>
   );
 };
 
@@ -96,55 +114,45 @@ const VizSettings = () => {
     e.preventDefault();
   };
   const handleChange = e => {
+    e.preventDefault();
     dispatch(e);
   };
   const onClick = () => {
     localStorage.setItem("cohendState", JSON.stringify(vizState));
   };
-  console.log("vis settings")
   return (
     <div>
       <Container maxWidth="sm" className={classes.container}>
         <Typography align="center" variant="h6" component="h3">
           Colors
         </Typography>
-        <Grid container justify="space-around">
-          <Grid item>
-            <Typography align="center" variant="caption" component="p">
+        <Grid container justify="space-around" alignItems="center">
+          <Grid xs={12} sm={4} item align="center">
+            <Typography variant="caption" component="p">
               Dist1
             </Typography>
             <ColorPicker dist="Dist1" />
           </Grid>
-          <Grid item>
-            <Typography align="center" variant="caption" component="p">
+          <Grid xs={12} sm={4} item align="center">
+            <Typography variant="caption" component="p">
               Overlap
             </Typography>
             <ColorPicker dist="DistOverlap" />
           </Grid>
-          <Grid item>
-            <Typography align="center" variant="caption" component="p">
+          <Grid xs={12} sm={4} item align="center">
+            <Typography variant="caption" component="p">
               Dist2
             </Typography>
             <ColorPicker dist="Dist2" />
           </Grid>
-          <Typography align="center" variant="caption" component="p">
+        </Grid>
+        <Typography align="center" variant="caption" component="p">
             Click to change colors.
           </Typography>
-          
-        </Grid>
         <Divider />
-        {/* <FormControlLabel
-        control={
-          <Switch
-            checked={state.darkMode}
-            onChange={() => dispatch({name: "toggleDarkMode"})}
-            name="dark mode"
-            color="primary"
-          />
-        }
-        label="Toggle Dark Mode"
-        labelPlacement="start"
-      /> */}
+        <Typography align="center" variant="body2" component="p">
+            Dark mode:     <DarkModeToggle />
+          </Typography>
         <Divider />
         <Typography align="center" variant="h6" component="h3">
           Parameters
