+++ rpsych-cohend/src/pages/cohend.js
@@ -1,4 +1,6 @@
 import React from 'react';
 import App from '../App';
 
-export default () => <App />
+export default (props) => {
+return (<App {...props} />)
+}
